Extract workspace generation runner hook
